refactor: 랭크 뷰모델 통합 및 뷰 코드 수정 #12
