+++ TheGee/TheGee/Sources/Presentation/GameView/Views/SpeedGameView.swift
@@ -14,7 +14,7 @@ struct SpeedGameView: View {
     // MARK: 프로퍼티
     @Environment(\.dismiss) private var dismissAction
     @StateObject private var viewModel = SpeedGameViewModel()
-    @StateObject private var personalRecordViewModel = PersonalRecordViewModel()
+    @StateObject private var rankViewModel = RankViewModel()
     
     // MARK: 바디
     var body: some View {
@@ -66,7 +66,7 @@ struct SpeedGameView: View {
                             viewModel.startGame()
                         },
                         onSaveRecord: {
-                            personalRecordViewModel.saveRecord(reactionTime: viewModel.calculateAverageReactionTime())
+                            rankViewModel.saveRecord(reactionTime: viewModel.calculateAverageReactionTime())
                         }
                     )
                     .transition(.opacity)
